.bin/weechat-mqtt-notify/requirements.txt
ansible==2.0.2.0
apsw==3.8.11.1.post1
atomicwrites==1.0.0
backports.shutil-get-terminal-size==1.0.0
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
cffi==1.6.0
chardet==2.3.0
CherryPy==3.5.0
click==6.6
click-log==0.1.4
click-threading==0.1.2
configobj==5.0.6
cryptography==1.3.1
cssselect==0.9.1
cssutils==1.0
dnspython==1.12.0
ecdsa==0.13
enum34==1.1.5
feedparser==5.1.3
html5lib==0.999
http==0.2
httplib2==0.9.1
idna==2.1
imaplib2==2.53
ipaddress==1.0.16
Jinja2==2.8
khard==0.9.0
lxml==3.6.0
Markdown==2.6.6
MarkupSafe==0.23
mechanize==0.2.5
ndg-httpsclient==0.4.0
netaddr==0.7.18
netifaces==0.10.4
notify2==0.3
numpy==1.11.0
paho-mqtt==1.1
paramiko==1.16.0
pathlib==1.0.1
pew==0.1.19
Pillow==3.2.0
powerline-status===dev
py-bcrypt==0.4
pyasn1==0.1.9
pyCardDAV==0.7.0
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
Pygments==2.1.3
pygobject==3.20.1
pylibacl==0.5.2
pyOpenSSL==16.0.0
pyparsing==2.1.1
PySocks==1.5.6
pysqlite==1.0.1
python-apt==1.1.0b2
python-dateutil==2.4.2
python-debian==0.1.27
python-debianbts==2.6.0
pythonz-bd==1.11.2
pyxattr==0.5.3
pyxdg==0.25
PyYAML==3.11
rdiff-backup==1.2.8
redis==2.10.5
reportbug==6.6.6
repoze.lru==0.6
requests==2.9.1
requests-toolbelt==0.6.2
Routes==2.2
rubber==1.4
shutilwhich==1.1.0
six==1.10.0
urllib3==1.13.1
urwid==1.3.1
uTidylib==0.2
virtualenv==15.0.1
virtualenv-clone==0.2.6
vobject==0.9.2
WebOb==1.6.0
wicd==1.7.4
